[AMDGPU] Add IR lowering changes for preloaded kernargs
Cherry-pick...

Preloaded kernel arguments should not be lowered in the IR pass
AMDGPULowerKernelArguments. Therefore it's necessary to calculate the
total number of user SGPRs that are available for preloading and how
many SGPRs would be required to preload each argument to determine
whether we should skip lowering i.e. the argument will be preloaded
instead.

Differential Revision: https://reviews.llvm.org/D156853